+++ R/read.R
@@ -14,6 +14,10 @@
 #'   per second. `nevents` is the number of recorded events.
 #' * **parameters**: A list with meta data of the recording. The parameters are
 #'   organized in groups, similarly to the original structure in the c3d file.
+#'   c3dr tries to preserve the data type and data structure of each imported
+#'   parameter value. One-dimensional parameters are imported as a scalar or
+#'   vector, two-dimensional parameters are imported as a matrix,
+#'   three-dimensional parameters are imported as a list of matrices.
 #' * **data**: A list with the point data of the recording. Each element in the
 #'   list corresponds to one frame. Use [c3d_data()] to convert the data to a
 #'   data frame.