+++ src/ezc3d/Analogs.h
@@ -9,6 +9,7 @@
 ///
 
 #include "ezc3d/AnalogsSubframe.h"
+#include <cstdint>
 
 ///
 /// \brief Analog holder for C3D analogous data