+++ NEWS.md
@@ -1,3 +1,23 @@
+c3dr (development version)
+===========================
+
+### NEW FEATURES
+
+  * Support export of force platform data in c3d_write(). Thanks to Michael Schwartz for suggesting and testing this feature.
+  
+### MINOR IMPROVEMENTS
+
+  * Preserve matrix data structure of imported parameters
+  * Implement latest changes to EZC3D
+
+c3dr 0.1.5 (2025-05-21)
+===========================
+
+### BUG FIXES
+
+  * Fix compilation error for gcc15
+
+
 c3dr 0.1.4 (2025-05-15)
 ===========================
 
@@ -5,7 +25,7 @@ First CRAN release.
 
 ### MINOR IMPROVEMENTS
 
-* Minor fixes for the CRAN submission
+  * Minor fixes for the CRAN submission
 
 c3dr 0.1.1 (2025-04-24)
 ========================